Files for RESTful web service
In this project, I have implemented a RESTful web service that utilises the appropriate HTTP methods (GET, POST, PUT and DELETE) to enable full CRUD functionality on the film database. The RESTful service should implement some mechanism to allow the client to choose which format the data should be returned in (i.e. JSON, XML and TEXT format). For the second part of the project, The web application uses the RESTful service that I have implemented to provide a cohesive user experience and present the film data to the user. The web app should allow users to browse and search the film database (search by title, year and genre). Furthermore, there should be mechanisms that facilitate inserting, updating and deleting films. Data exchange between client and server is done through JSON, XML or Text/String formats. Meaning the web app should request data from the RESTful service in one of the specified formats (JSON, XML and TEXT), the web app should also send data to the server in one of these formats.